Directly publish diagnostics and rate-limit connection attempts
The diagnostics were previously published using the diagnostic_updater
class, but the current implementation doesn't align well with lifecycle
nodes. The diagnostics published by this driver aren't complex, so it is
simple to just craft the message manually.

When the device goes into a state where re-connection is attempted, it
can tie up CPU resources if the connection attempts are performed
quickly. This change limits connection attempts to 5 Hz.